handle validation errors
